requirements.txt
atomic==0.7.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
